Prevent double or hidden legend items from showing image
